+++ packages/web/src/views/todos/index.vue
@@ -1,8 +1,10 @@
 <script setup lang="ts">
 import TodoList from '@/components/TodoList.vue'
+import TodoForm from '@/components/TodoForm.vue'
 </script>
 
 <template lang="pug">
 main
   TodoList
+  TodoForm
 </template>
